Move func to a separate module
